backend/requirements.txt
# FastAPI and Web Server
fastapi==0.104.1
uvicorn[standard]==0.24.0
websockets==12.0
python-multipart==0.0.6

# Database
sqlalchemy==2.0.23
psycopg2-binary>=2.9.9
alembic==1.12.1

# Environment Variables
python-dotenv==1.0.0
pydantic==2.5.0
pydantic-settings==2.1.0

# Image Processing (will be used in Phase 2)
opencv-python==4.8.1.78
numpy==1.24.3
Pillow==10.1.0

# Face Recognition (will be used in Phase 2)
face-recognition==1.3.0
dlib==19.24.2

# Utilities
python-dateutil==2.8.2

# Authentication
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4

# Caching
redis>=5.0.0
hiredis>=2.2.3  # Faster Redis parser

# Testing (optional, for test scripts)
# requests>=2.31.0  # For API testing
# websockets>=12.0  # For WebSocket testing (already included above)

